requirements.txt
# Django and REST framework
Django==5.0.4
djangorestframework==3.14.0
django-cors-headers==4.7.0
django-allauth==65.4.1
dj-rest-auth==7.0.1
djangorestframework_simplejwt==5.4.0

# Database
mysql-connector-python==8.0.33
psycopg2-binary==2.9.10

# File Processing
PyPDF2==3.0.1
Pillow==11.1.0
python-magic==0.4.27

# API Documentation
drf-yasg==1.21.7
Markdown==3.7

# Security and Authentication
# python-dotenv==1.0.1
# python-decouple==3.8
# cryptography==44.0.0

# HTTP and Requests
requests==2.31.0
urllib3==2.2.1
aiohttp==3.9.3

# Date/Time
python-dateutil==2.9.0.post0
pytz==2024.1

# Utils
python-json-logger==3.2.1
PyYAML==6.0.2
whitenoise==6.6.0

# AI/ML Related
google-generativeai==0.8.3
# nltk==3.8.1
# numpy==1.24.4
# pandas==1.5.3

# Frontend Integration
django-tailwind==3.8.0

langdetect==1.0.9

gunicorn==21.2.0
